Update mio to 0.6
The main differences between mio 0.5 and mio 0.6 is that the Handler
trait goes away, EventLoop goes away, and a new Poll type is introduced
which makes the API similar to epoll. Naturally, this required some API
reorganization. One other side effect of the update is that mio
consumers are now responsible for managing timers and channels
externally.

The event loop in src/watch was replaced with a simple mpsc channel
since there was no hard requirement for mio.

Also added some logging for debugging.