change connect log messages from info to trace level
Currently info level log messages are displayed when connections
are established, it would be better to have these as trace level
especially when considering the actual log message.

Changes the log macro from info to trace to reflect the desired log
level for these messages.

Logging output is now quieter.